moved chat links to left bar
