+++ src/ChatList.jsx
@@ -65,18 +65,29 @@ export const ChatList = (props) => {
   		let name = chat.name;
       let encodedName = encodeURIComponent(name); // This variable is safe for use in a URL.
   		let id = chat.id;
+      /*
   		return <tr key={id}>
   				  <td><Link className="link" to={`${encodedName}/${id}`}>{name}</Link></td>
   				  <td>{users}</td>
             <td><button onClick={() => handleClick(name, id)}>x</button></td>
   			   </tr>
+      */
+      return <li className="chat-item">
+              <button className="remove-btn" id="remove" onClick={() => handleClick(name, id)}>X</button>
+              <div className="chat-info">
+                <Link className="chat-link" to={`${encodedName}/${id}`}><b>{name}</b>
+                  <p className="ptcp-list">{users}</p>
+                </Link>
+              </div>
+            </li>
+
+
   	})
     return(chatList);
   }
 
-	return(
-    <div>
-      <div className="bottom-rounded-box" id="my-chats">
+  /*
+        <div className="bottom-rounded-box" id="my-chats">
         <h3 className="title">My chat rooms</h3>
         <div>
           <table className="chat-table">
@@ -91,6 +102,13 @@ export const ChatList = (props) => {
           </table>
         </div>
       </div>
+  */
+
+	return(
+    <div>
+      <ul className="chat-list">
+        {displayChats()}
+      </ul>
     </div>
 	 );
 }
