+++ src/StartChat.jsx
@@ -83,13 +83,7 @@ export class StartChat extends React.Component {
 
     const { username, chats, participants } = this.props;
   	return(
-  	  <div className="main-container">
-        <div className="left-bar">
-          <div className="chats-title-container">
-            <h3 className="chats-title">My chats</h3>
-            <button class="add-chat-btn">+</button>
-          </div>
-        </div>
+  	  <div>
         <div className="top-rounded-box" id="start-chat">
           <h3 className="title">Start a new chat</h3>
           <div id="chat-form-container">
@@ -105,7 +99,6 @@ export class StartChat extends React.Component {
             </form>
           </div>
         </div>
-        <ChatList username={username} chats={chats} participants={participants} removeFromChat={this.props.removeFromChat} />
       </div>
     );
   }
